requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-filters==0.11.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
python-telegram-bot==22.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
